+++ test/unit/mysql_rep_admin/run_program.py
@@ -54,6 +54,57 @@ def rpt_slv_log(master, slaves, form, ofile, db_tbl, class_cfg):
     return True
 
 
+class SlaveRep(object):
+
+    """Class:  SlaveRep
+
+    Description:  Class stub holder for mysql_class.SlaveRep class.
+
+    Super-Class:  None
+
+    Sub-Classes:  None
+
+    Methods:
+        __init__ -> Class initialization.
+        connect -> Stub method holder for SlaveRep.connect.
+
+    """
+
+    def __init__(self, name=None, sid=None, user=None, passwd=None,
+                 serv_os=None, host=None, port=None, cfg_file=None):
+
+        """Method:  __init__
+
+        Description:  Class initialization.
+
+        Arguments:
+            name -> Stub holder.
+            sid -> Stub holder.
+            user -> Stub holder.
+            passwd -> Stub holder.
+            serv_os -> Stub holder.
+            host -> Stub holder.
+            port -> Stub holder.
+            cfg_file -> Stub holder.
+
+        """
+
+        self.name = name
+        self.conn = True
+
+    def connect(self):
+
+        """Method:  connect
+
+        Description:  Stub method holder for SlaveRep.connect.
+
+        Arguments:
+
+        """
+
+        return True
+
+
 class MasterRep(object):
 
     """Class:  MasterRep
@@ -180,13 +231,14 @@ class UnitTest(unittest.TestCase):
                            "user": "root", "serv_os": "Linux", "sid": "21",
                            "port": "3306"}]
 
+    @mock.patch("mysql_rep_admin.mysql_class.SlaveRep")
     @mock.patch("mysql_rep_admin.mysql_class.MasterRep")
     @mock.patch("mysql_rep_admin.cmds_gen.disconnect")
     @mock.patch("mysql_rep_admin.call_run_chk")
     @mock.patch("mysql_rep_admin.cmds_gen.create_cfg_array")
     @mock.patch("mysql_rep_admin.gen_libs.load_module")
     def test_single_func(self, mock_cfg, mock_array, mock_call, mock_dis,
-                         mock_rep):
+                         mock_rep, mock_slv):
 
         """Function:  test_single_func
 
@@ -201,6 +253,7 @@ class UnitTest(unittest.TestCase):
         mock_call.return_value = True
         mock_dis.return_value = True
         mock_rep.return_value = MasterRep()
+        mock_slv.return_value = SlaveRep()
 
         self.assertFalse(mysql_rep_admin.run_program(self.args_array,
                                                       self.func_dict))