+++ mysql_rep_admin.py
@@ -288,6 +288,8 @@ def chk_slv_thr(MASTER, SLAVE, **kwargs):
 
     """
 
+    SLAVE = list(SLAVE)
+
     if SLAVE:
 
         for slv in SLAVE: